Efficient Inference/requirements.txt
# PyTorch index (choose the right one for your setup)
--extra-index-url https://download.pytorch.org/whl/cu121

# Core dependencies
torch>=2.0.0
diffusers>=0.30.0
git+https://github.com/huggingface/transformers.git
pillow>=9.0.0
numpy>=1.21.0
huggingface_hub>=0.16.0

# Monitoring dependencies
psutil>=5.9.0
nvidia-ml-py>=12.0.0  # pynvml for GPU monitoring
pandas>=1.5.0
matplotlib>=3.5.0

# Optional: For evaluation
easyocr>=1.7.0
python-Levenshtein
scikit-image
scipy
optimum-quanto